// frontend/apps/stitch-admin/src/shared/api/generated/model/createDestinationRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Stitch API Gateway
 * OpenAPI spec version: 1.0.0
 */
import type { AddressDTO } from './addressDTO';
import type { DestinationType } from './destinationType';

export interface CreateDestinationRequest {
  name: string;
  type: DestinationType;
  address: AddressDTO;
  contact_name?: string | null;
  contact_phone?: string | null;
}