// frontend/apps/stitch-admin/src/shared/api/generated/model/destinationType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Stitch API Gateway
 * OpenAPI spec version: 1.0.0
 */

export type DestinationType = typeof DestinationType[keyof typeof DestinationType];


export const DestinationType = {
  DDN_BASE: 'DDN_BASE',
  LANDFILL: 'LANDFILL',
  RECYCLING_CENTER: 'RECYCLING_CENTER',
  TREATMENT_PLANT: 'TREATMENT_PLANT',
  OTHER: 'OTHER',
} as const;